Split up health class
